Fix issue #522 : Transliterate Caller Name
