+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/model/DeviceService.java
@@ -60,6 +60,7 @@ public interface DeviceService extends EventHandler {
     String EXTRA_VIBRATION_INTENSITY = "vibration_intensity";
     String EXTRA_CALL_COMMAND = "call_command";
     String EXTRA_CALL_PHONENUMBER = "call_phonenumber";
+    String EXTRA_CALL_DISPLAYNAME = "call_displayname";
     String EXTRA_CANNEDMESSAGES = "cannedmessages";
     String EXTRA_CANNEDMESSAGES_TYPE = "cannedmessages_type";
     String EXTRA_MUSIC_ARTIST = "music_artist";
